+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 0.0.4 - TBD
+## 0.0.4 - 2024-12-10
 
 * Fixed bug where folding was not provided on an untitled/yet-to-be-saved editor even if the language was set to `daily-bullet-notes`
 * Added auto completion options when cursor is within a `[ ]` box